Fixes code that was nested too deeep. Refactors
